requirements.txt
asgiref==3.7.2
class-registry==2.1.2
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
regex==2023.8.8
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
django-tenants==3.5.0
